feat: improve generators
