+++ Sources/DesignTokensGenerator/Generators/SourceCodeGenerator.swift
@@ -1,0 +1,7 @@
+import Foundation
+import Stencil
+import StencilSwiftKit
+
+protocol SourceCodeGenerator {
+  func generate(with environment: Stencil.Environment) throws -> [SourceCodeFile]
+}
